refactor(CardBrowser): extract 'toggleMark'
change extracted from a large refactoring: 14843
